Subsetting/slicing/boolean indexing enhanced

+++ include/np/Constants.hpp
@@ -25,11 +25,11 @@ SOFTWARE.
 #pragma once
 
 #include <cmath>
-#include <cstddef>
+#include <cstdint>
 
 namespace np {
 
-    using Size = std::size_t;
+    using Size = std::uint32_t;
     static const Size constexpr SIZE_DEFAULT = static_cast<Size>(-1);
 
     static const double constexpr NaN = NAN;
